requirements.txt
ultralytics
torch
opencv-python
numpy
pillow
streamlit
